Adding flyway and jdbc for medicine and patient_medicine

+++ src/main/java/com/hashini/medicare/dao/PrescriptionDAO.java
@@ -14,5 +14,5 @@ public interface PrescriptionDAO {
 
     Optional<PrescriptionDTO> selectPrescriptionById(long id);
 
-    int addPrescription(Prescription prescription);
+    long addPrescription(Prescription prescription);
 }
